requirements-api.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
google-generativeai>=0.3.0
websockets==12.0
